+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 89.0.0
+
+* BREAKING: Change source for `GdsApi.worldwide.world_locations` method and remove pagination.
+* Change source for `GdsApi.worldwide.world_location` method (the response is backwards compatible).
+
 # 88.2.0
 
 * Add `whitehall_media` method to Asset Manager.